Template Istio version (#80)
We should avoid hard coding the Istio version in the Istio values.yaml
file. By templating the values we can resolve the Istio version and
consume new Istio releases by updating a single value.

This also updates the template format to evaluate expressions within
`{{ expr }}` instead of evaluating every line as a shell expression.
This makes escaping values that should not be matched much simpler as
only `{{ ... }}` will be evaluated.