Publish "uncharted" yaml along side charts (#87)
In addition to charts, we should publish the same source files as plain yaml. Images are still resolved to digests.

This will make it easier for non-helm based install via tools like kapp or kubectl.

Published uncharts are at:
> gs://projectriff/charts/uncharted/${version}/${chart}.yaml

Snapshots are at:
> gs://projectriff/charts/uncharted/snapshots/${slug}/${charts}.yaml

Also updating the riff system build which no longer has multiple riff-system namespaces, so we no longer need the overlays to strip the extra definitions. (The namespace is only defined in riff-build)

Refs #86